+++ Editor/Editors/TextureEditor/TextureEditorView.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows.Controls;
+using System.Windows.Input;
 
 namespace Editor.Editors
 {
@@ -10,6 +11,13 @@ namespace Editor.Editors
         public TextureEditorView()
         {
             InitializeComponent();
+            Focus();
         }
+
+        private void centerTexture_Executed(object sender, ExecutedRoutedEventArgs e) => textureView.Center();
+        private void zoomInTexture_Executed(object sender, ExecutedRoutedEventArgs e) => textureView.ZoomIn();
+        private void zoomOutTexture_Executed(object sender, ExecutedRoutedEventArgs e) => textureView.ZoomOut();
+        private void zoomFitTexture_Executed(object sender, ExecutedRoutedEventArgs e) => textureView.ZoomFit();
+        private void actualTextureSize_Executed(object sender, ExecutedRoutedEventArgs e) => textureView.ActualSize();
     }
 }